+++ src/app.module.js
@@ -1,0 +1,5 @@
+const app = angular.module('app', [
+  'ngRoute',
+  'ngAnimate',
+  'movieList'
+]);
